+++ utoipa-gen/src/component/schema/features.rs
@@ -64,7 +64,8 @@ impl Parse for UnnamedFieldStructFeatures {
             ContentEncoding,
             ContentMediaType,
             Bound,
-            NoRecursion
+            NoRecursion,
+            Pattern
         )))
     }
 }